Ajout de l'écran de mort + retour au menu
(Appuyer temporairement sur Suppr pour tuer le joueur)

+++ Configuration.pde
@@ -8,7 +8,7 @@
 final int LARGEUR_ECRAN = 320;
 final int HAUTEUR_ECRAN = 180;
 final int ECHELLE = 3;
-final int IMAGES_PAR_SECONDE = 60;
+final float IMAGES_PAR_SECONDE = 60;
 ;
 
 /* Images */
